[#24] refactor: 환경변수 추가, 충돌 오류 해결

+++ src/main/java/com/labhub/CveLabhubBack/news/repository/CveNewsMappingRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface CveNewsMappingRepository extends JpaRepository<CveNewsMapping, CveNewsMapping.CveNewsMappingId> {
 
-    boolean existsByNewsIdAndCveId(Long newsId, Long cveId);
+    boolean existsByNewsIdAndCveId(Long newsId, Integer cveId);
 }
